refactor daemon peer credential tests
