--- cmd/agent-compose/peercred_darwin.go
@@ -1,31 +1,0 @@
-//go:build darwin
-
-package main
-
-import (
-	"net"
-
-	"golang.org/x/sys/unix"
-)
-
-// unixSocketPeerUID returns the effective UID of the process on the other end
-// of the Unix socket connection.
-func unixSocketPeerUID(conn *net.UnixConn) (int, error) {
-	raw, err := conn.SyscallConn()
-	if err != nil {
-		return 0, err
-	}
-	var (
-		cred    *unix.Xucred
-		credErr error
-	)
-	if err := raw.Control(func(fd uintptr) {
-		cred, credErr = unix.GetsockoptXucred(int(fd), unix.SOL_LOCAL, unix.LOCAL_PEERCRED)
-	}); err != nil {
-		return 0, err
-	}
-	if credErr != nil {
-		return 0, credErr
-	}
-	return int(cred.Uid), nil
-}
